% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notcore_utils.r
\name{csmf_eava}
\alias{csmf_eava}
\title{Calculate CSMF from EAVA::codEAVA output}
\usage{
csmf_eava(x)
}
\arguments{
\item{x}{output from EAVA::codEAVA}
}
\value{
Named vector
}
\description{
Calculate CSMF from EAVA::codEAVA output
}
