% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DataEAVA}
\alias{DataEAVA}
\title{1,736 records of Sample Input}
\format{
1,736 arbitrary input records.
}
\description{
This is a dataset of 1,736 arbitrary (neonate and child) deaths in the
acceptable format of EAVA.  It is meant to be comparable
to the RandomVA6 examples data set.
}
\examples{

data(DataEAVA)

}
\keyword{datasets}
