% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grouping_eava_child}
\alias{grouping_eava_child}
\title{Mapping for child causes of death to the list used by the EAVA algorithm.}
\format{
Data Frame
}
\description{
Intended for use when comparing assigned causes of death between
the EAVA algorithm and InSilicoVA/InterVA5.
}
\details{
#'
}
\examples{

data(grouping_eava_child)

}
\keyword{datasets}
