% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NeonatesVA5}
\alias{NeonatesVA5}
\title{200 records of Sample Input}
\format{
200 arbitrary input records.
}
\description{
This is a dataset consisting of 200 arbitrary sample input neonatal deaths 
in the acceptable format of InterVA5.
}
\examples{

data(NeonatesVA5)

}
\keyword{datasets}
