% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.r
\name{openVA_status}
\alias{openVA_status}
\title{Check openVA packages status}
\usage{
openVA_status()
}
\description{
This will print the current versions of all openVA packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
openVA_status()
}
}
\seealso{
Other package status: 
\code{\link{openVA_update}()}
}
\concept{package status}
