% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.r
\name{openVA_update}
\alias{openVA_update}
\title{Update openVA packages}
\usage{
openVA_update()
}
\description{
This will check to see if all openVA packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
openVA_update()
}
}
\seealso{
Other package status: 
\code{\link{openVA_status}()}
}
\concept{package status}
