% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notcore_utils.r
\name{plot.eava}
\alias{plot.eava}
\title{Create CSMF plot for EAVA::codEAVA output}
\usage{
\method{plot}{eava}(
  x,
  top = 10,
  title = "Top CSMF Distribution",
  type = "bar",
  return.barplot = FALSE,
  ...
)
}
\arguments{
\item{x}{Output from EAVA::codEAVA}

\item{top}{The number of top causes to include in plot. This is exceeded if
there are ties.}

\item{title}{Title for CSMF plot}

\item{type}{An indicator of the type of chart to plot.  "pie" for pie chart;
"bar" for bar chart.}

\item{return.barplot}{a logical indicating if the (barplot) ggplot() object
should be returned (instead of printed).  Default value is FALSE.}

\item{\dots}{Not used.}
}
\value{
A barplot if return.barplot is TRUE; otherwise, nothing is returned.
}
\description{
Create CSMF plot for EAVA::codEAVA output
}
