% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepCalibration.r
\name{prepCalibration}
\alias{prepCalibration}
\title{Convert the output from codeVA to the format expected by
vacalibration::vacalibration().}
\usage{
prepCalibration(fit, ...)
}
\arguments{
\item{fit}{a fitted object from \code{codeVA} using either InSilicoVA or
InterVA5}

\item{...}{more fitted objects from \code{codeVA} using either InSilicoVA or
InterVA5}
}
\value{
a list with one data.frame that contains variables "ID" and "cause",
and the element's name (in the list) is the algorithm name used to assign
causes (either "insilicova" or "interva")
}
\description{
Convert the output from codeVA to the format expected by
vacalibration::vacalibration().
}
\examples{
\dontrun{
data(NeonatesVA5)
fit_insilico <- codeVA(NeonatesVA5)
insilico_prep <- prepCalibration(fit_insilico)
calib_insilico = vacalibration::vacalibration(va_data = insilico_prep,
                                              age_group = "neonate",
                                              country = "Mozambique")
round(calib_insilico$pcalib_postsumm["insilicova",,], 3)

fit_interva <- codeVA(NeonatesVA5, model = "InterVA", version = "5", write = FALSE)
interva_prep <- prepCalibration(fit_interva)
calib_interva = vacalibration::vacalibration(va_data = interva_prep,
                                             age_group = "neonate",
                                             country = "Mozambique")
round(calib_interva$pcalib_postsumm["interva",,], 3)

two_fits <- prepCalibration(fit_insilico, fit_interva)
calib_out_ensemble = vacalibration::vacalibration(va_data = two_fits,
                                                  age_group = "neonate",
                                                  country = "Mozambique")
round(calib_out_ensemble$pcalib_postsumm["ensemble",,], 3)

} 
}
