% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notcore_utils.r
\name{print.eava}
\alias{print.eava}
\title{Print method for "eava" class.}
\usage{
\method{print}{eava}(x)
}
\arguments{
\item{x}{\code{eava} object.

\dontrun{
data(DataEAVA)
eava_results <- codeVA(DataEAVA, data.type = "EAVA", model = "EAVA")
eava_results
}}
}
\description{
Print method for "eava" class.
}
