% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notcore_utils.r
\name{print.eava_summary}
\alias{print.eava_summary}
\title{Print method for summarizing results from EAVA algorithm.}
\usage{
\method{print}{eava_summary}(x, ...)
}
\arguments{
\item{x}{\code{eava} object}

\item{\dots}{not used}
}
\description{
This function prints a summary message of the results along with
the top cause-specific mortality fractions (CSMFs).
}
