% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notcore_utils.r
\name{print.vacalibration}
\alias{print.vacalibration}
\title{Print method for vacalibration model fits}
\usage{
\method{print}{vacalibration}(x, ...)
}
\arguments{
\item{x}{\code{vacalibration} object}

\item{...}{not used}
}
\description{
This function is the print method for class \code{vacalibration}
}
\examples{
\dontrun{
data(NeonatesVA5)
fit_insilico <- codeVA(NeonatesVA5, auto.length = FALSE)
insilico_prep <- prepCalibration(fit_insilico)
calib_insilico = vacalibration::vacalibration(va_data = insilico_prep,
                                              age_group = "neonate",
                                              country = "Mozambique",
                                              plot_it = FALSE)
calib_insilico
}
}
