% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RandomVA6}
\alias{RandomVA6}
\title{1,736 records of Sample Input}
\format{
1,736 arbitrary input records.
}
\description{
This is a dataset consisting of 1,736 (neonate and child) deaths in
the acceptable format of InSilicoVA.  It is meant to be comparable
to the DataEAVA examples data set.
}
\examples{

data(RandomVA6)

}
\keyword{datasets}
