% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{insilico}
\alias{InterVA5}
\alias{tariff}
\alias{InterVA}
\alias{CSMF}
\alias{CSMF5}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{InSilicoVA}{\code{\link[InSilicoVA]{insilico}}}

  \item{InterVA4}{\code{\link[InterVA4]{CSMF}}, \code{\link[InterVA4]{InterVA}}}

  \item{InterVA5}{\code{\link[InterVA5]{CSMF5}}, \code{\link[InterVA5]{InterVA5}}}

  \item{Tariff}{\code{\link[Tariff]{tariff}}}
}}

