% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notcore_utils.r
\name{summary.eava}
\alias{summary.eava}
\title{Summary of results obtained by fitting the EAVA algorithm.}
\usage{
\method{summary}{eava}(x, top = 5, rnd = 4)
}
\arguments{
\item{x}{\code{eava} object}

\item{top}{number of top CSMF to show}

\item{rnd}{number of decimal places to round the CSMF}
}
\description{
This function prints a summary message of the results along with
the top cause-specific mortality fractions (CSMFs).
}
