% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ConvertData.r
\name{getPHMRC_url}
\alias{getPHMRC_url}
\title{Get the URL to the PHMRC dataset}
\usage{
getPHMRC_url(type)
}
\arguments{
\item{type}{adult, child, or neonate}
}
\value{
URL of the corresponding dataset
}
\description{
Get the URL to the PHMRC dataset
}
\examples{
link <- getPHMRC_url("adult")
summary(link)$description
}

