% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VAmain.r
\name{plotVA}
\alias{plotVA}
\title{Plot top CSMF for a fitted model}
\usage{
plotVA(object, top = 10, title = NULL, ...)
}
\arguments{
\item{object}{a fitted object using \code{\link{codeVA}}}

\item{top}{number of top causes to plot}

\item{title}{title of the plot}

\item{...}{additional arguments passed to \code{\link[InSilicoVA]{plot.insilico}},
\code{\link[Tariff]{plot.tariff}}, \code{\link[InterVA4]{CSMF}}, or \code{\link[nbc4va]{plot.nbc}}.}
}
\description{
Plot top CSMF for a fitted model
}
\examples{
\donttest{
data(RandomVA3)
test <- RandomVA3[1:200, ]
train <- RandomVA3[201:400, ]
fit1 <- codeVA(data = test, data.type = "customize", model = "InSilicoVA",
                    data.train = train, causes.train = "cause",
                    Nsim=1000, auto.length = FALSE)

fit2 <- codeVA(data = test, data.type = "customize", model = "InterVA",
               data.train = train, causes.train = "cause",
               version = "4.02", HIV = "h", Malaria = "l")

fit3 <- codeVA(data = test, data.type = "customize", model = "Tariff",
               data.train = train, causes.train = "cause",
               nboot.sig = 100)

fit4 <- codeVA(data = test, data.type = "customize", model = "NBC",
               data.train = train, causes.train = "cause", known.nbc = TRUE)

plotVA(fit1)
plotVA(fit2)
plotVA(fit3)
plotVA(fit4)
}
}
\seealso{
\code{\link[InSilicoVA]{plot.insilico}}, \code{\link[Tariff]{plot.tariff}},
\code{\link[InterVA4]{CSMF}}
}

